+++ tests/cache/api.twitch.tv.v5-oembed.825b0cd75e46697eb1798bc6f1b06547.php
@@ -1,0 +1,52 @@
+<?php
+declare(strict_types = 1);
+
+return [
+    'headers' => [
+        'access-control-allow-origin' => [
+            '*'
+        ],
+        'cache-control' => [
+            'no-cache, no-store, must-revalidate, private'
+        ],
+        'expires' => [
+            '0'
+        ],
+        'pragma' => [
+            'no-cache'
+        ],
+        'timing-allow-origin' => [
+            'https://www.twitch.tv'
+        ],
+        'date' => [
+            'Sat, 23 May 2020 23:02:25 GMT'
+        ],
+        'x-served-by' => [
+            'cache-sea4472-SEA, cache-mad22046-MAD'
+        ],
+        'x-cache' => [
+            'MISS, MISS'
+        ],
+        'x-cache-hits' => [
+            '0, 0'
+        ],
+        'x-timer' => [
+            'S1590274945.952174,VS0,VS0,VE158'
+        ],
+        'vary' => [
+            'Accept-Encoding'
+        ],
+        'strict-transport-security' => [
+            'max-age=300'
+        ],
+        'Content-Location' => [
+            'https://api.twitch.tv/v5/oembed?url=https%3A%2F%2Fwww.twitch.tv%2Ftwit&format=json'
+        ],
+        'X-Request-Time' => [
+            '0.252 ms'
+        ]
+    ],
+    'statusCode' => 204,
+    'reasonPhrase' => 'No Content',
+    'body' => ''
+];