+++ tests/cache/api.twitch.tv.v5-oembed.b8bbd0b88460a8e79657ac7c373cadce.php
@@ -1,0 +1,52 @@
+<?php
+declare(strict_types = 1);
+
+return [
+    'headers' => [
+        'access-control-allow-origin' => [
+            '*'
+        ],
+        'cache-control' => [
+            'no-cache, no-store, must-revalidate, private'
+        ],
+        'expires' => [
+            '0'
+        ],
+        'pragma' => [
+            'no-cache'
+        ],
+        'timing-allow-origin' => [
+            'https://www.twitch.tv'
+        ],
+        'date' => [
+            'Sat, 23 May 2020 23:02:24 GMT'
+        ],
+        'x-served-by' => [
+            'cache-sea4482-SEA, cache-mad22027-MAD'
+        ],
+        'x-cache' => [
+            'MISS, MISS'
+        ],
+        'x-cache-hits' => [
+            '0, 0'
+        ],
+        'x-timer' => [
+            'S1590274945.528781,VS0,VS0,VE159'
+        ],
+        'vary' => [
+            'Accept-Encoding'
+        ],
+        'strict-transport-security' => [
+            'max-age=300'
+        ],
+        'Content-Location' => [
+            'https://api.twitch.tv/v5/oembed?url=https%3A%2F%2Fwww.twitch.tv%2Fvideos%2F106400740&format=json'
+        ],
+        'X-Request-Time' => [
+            '0.261 ms'
+        ]
+    ],
+    'statusCode' => 204,
+    'reasonPhrase' => 'No Content',
+    'body' => ''
+];